+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 25.4.6
 * Improved content error handling and display mechanics.
+* Updated user properties caching mechanism according to sessions.
 
 ## 25.4.5
 * Added a new config flag `setUseSerialExecutor(boolean useSerial)` for selecting immediate request executor type.